Deployment/uxo_detection_api/requirements.txt
flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
ultralytics==8.3.107
pillow==11.2.1
torch==2.6.0
torchvision==0.21.0
numpy==1.23.5